services/composio_auth/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography>=41.0.0

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Data Validation - using compatible versions
pydantic>=2.6.4,<3.0.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# CORS and Middleware
fastapi-cors==0.0.6

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# External Service Integrations
openai>=1.3.7
composio>=0.8.8
redis==5.0.1

# Additional dependencies
python-multipart==0.0.19
requests==2.32.5
websockets==15.0.1
tenacity